[#159530254] - Refactor/fix status update on lease up apps table
