+++ app/javascript/components/lease_ups/LeaseUpApplicationsPage.js
@@ -1,5 +1,6 @@
 import React from 'react'
-import { flow, map } from 'lodash'
+import { flow, map, each, set, clone } from 'lodash'
+import moment from 'moment'
 
 import LeaseUpApplicationsTableContainer from './LeaseUpApplicationsTableContainer'
 import TableLayout from '../layouts/TableLayout'
@@ -10,11 +11,23 @@ import apiService from '~/apiService'
 import { EagerPagination, SERVER_PAGE_SIZE } from '~/utils/EagerPagination'
 import Context from './context'
 import mapProps from '~/utils/mapProps'
+import utils from '~/utils/utils'
 
 const ROWS_PER_PAGE = 5
 
 class LeaseUpApplicationsPage extends React.Component {
-  state = { loading: false, applications: [], pages: 0 }
+  state = {
+    loading: false,
+    applications: [],
+    pages: 0,
+    statusModal: {
+      isOpen: false,
+      status: null,
+      applicationId: null,
+      showAlert: null,
+      loading: false
+    }
+  }
 
   constructor (props) {
     super(props)
@@ -40,6 +53,53 @@ class LeaseUpApplicationsPage extends React.Component {
     this.loadPage(state.page)
   }
 
+  handleCreateStatusUpdate = async (data) => {
+    const { applicationId, status } = data
+    const { applications } = this.state
+    const response = await apiService.createFieldUpdateComment(data)
+
+    if (response) {
+      each(applications, (app, index) => {
+        if (app.application_id === applicationId) {
+          this.updateResults(`[${index}]['lease_up_status']`, status)
+          this.updateResults(`[${index}]['status_updated']`, moment().format(utils.SALESFORCE_DATE_FORMAT))
+        }
+      })
+
+      this.updateStatusModal({
+        applicationId: null,
+        isOpen: false,
+        loading: false,
+        showAlert: false,
+        status: null
+      })
+    } else {
+      this.updateStatusModal({
+        loading: false,
+        showAlert: true
+      })
+    }
+  }
+
+  updateStatusModal = (values) => {
+    this.setState(prevState => {
+      return {
+        statusModal: {
+          ...clone(prevState.statusModal),
+          ...values
+        }
+      }
+    })
+  }
+
+  updateResults = (path, value) => {
+    this.setState(prevState => {
+      return {
+        applications: set(clone(prevState.applications), path, value)
+      }
+    })
+  }
+
   render () {
     const listing = this.props.listing
 
@@ -62,7 +122,10 @@ class LeaseUpApplicationsPage extends React.Component {
       handleOnFetchData: this.handleOnFetchData,
       loading: this.state.loading,
       pages: this.state.pages,
-      rowsPerPage: ROWS_PER_PAGE
+      rowsPerPage: ROWS_PER_PAGE,
+      handleCreateStatusUpdate: this.handleCreateStatusUpdate,
+      updateStatusModal: this.updateStatusModal,
+      statusModal: this.state.statusModal
     }
 
     return (
